remove 7 unnecessary #include "ptp-pack.c" statements
Move the byte-order-swapping helpers from ptp-pack.c to ptp.h as well as 2
ptp_free_deviceinfo variants and remove all but one ptp-pack.c inclusions.

This speeds up compile time and removes lots of compiler warning
repetitions if one shows up in ptp-pack.c.
