+++ camlibs/ptp2/ptp-pack.c
@@ -41,110 +41,6 @@
 # define MIN(a, b) ((a) < (b) ? (a) : (b))
 #endif
 
-static inline uint16_t
-htod16p (PTPParams *params, uint16_t var)
-{
-	return ((params->byteorder==PTP_DL_LE)?htole16(var):htobe16(var));
-}
-
-static inline uint32_t
-htod32p (PTPParams *params, uint32_t var)
-{
-	return ((params->byteorder==PTP_DL_LE)?htole32(var):htobe32(var));
-}
-
-static inline void
-htod16ap (PTPParams *params, unsigned char *a, uint16_t val)
-{
-	if (params->byteorder==PTP_DL_LE)
-		htole16a(a,val);
-	else
-		htobe16a(a,val);
-}
-
-static inline void
-htod32ap (PTPParams *params, unsigned char *a, uint32_t val)
-{
-	if (params->byteorder==PTP_DL_LE)
-		htole32a(a,val);
-	else
-		htobe32a(a,val);
-}
-
-static inline void
-htod64ap (PTPParams *params, unsigned char *a, uint64_t val)
-{
-	if (params->byteorder==PTP_DL_LE)
-		htole64a(a,val);
-	else
-		htobe64a(a,val);
-}
-
-static inline uint16_t
-dtoh16p (PTPParams *params, uint16_t var)
-{
-	return ((params->byteorder==PTP_DL_LE)?le16toh(var):be16toh(var));
-}
-
-static inline uint32_t
-dtoh32p (PTPParams *params, uint32_t var)
-{
-	return ((params->byteorder==PTP_DL_LE)?le32toh(var):be32toh(var));
-}
-
-static inline uint64_t
-dtoh64p (PTPParams *params, uint64_t var)
-{
-	return ((params->byteorder==PTP_DL_LE)?le64toh(var):be64toh(var));
-}
-
-static inline uint16_t
-dtoh16ap (PTPParams *params, const unsigned char *a)
-{
-	return ((params->byteorder==PTP_DL_LE)?le16atoh(a):be16atoh(a));
-}
-
-static inline uint32_t
-dtoh32ap (PTPParams *params, const unsigned char *a)
-{
-	return ((params->byteorder==PTP_DL_LE)?le32atoh(a):be32atoh(a));
-}
-
-static inline uint64_t
-dtoh64ap (PTPParams *params, const unsigned char *a)
-{
-	return ((params->byteorder==PTP_DL_LE)?le64atoh(a):be64atoh(a));
-}
-
-#define htod8a(a,x)	*(uint8_t*)(a) = x
-#define htod16a(a,x)	htod16ap(params,a,x)
-#define htod32a(a,x)	htod32ap(params,a,x)
-#define htod64a(a,x)	htod64ap(params,a,x)
-#define htod16(x)	htod16p(params,x)
-#define htod32(x)	htod32p(params,x)
-#define htod64(x)	htod64p(params,x)
-
-#define dtoh8a(x)	(*(uint8_t*)(x))
-#define dtoh16a(a)	dtoh16ap(params,a)
-#define dtoh32a(a)	dtoh32ap(params,a)
-#define dtoh64a(a)	dtoh64ap(params,a)
-#define dtoh16(x)	dtoh16p(params,x)
-#define dtoh32(x)	dtoh32p(params,x)
-#define dtoh64(x)	dtoh64p(params,x)
-
-static inline uint32_t _post_inc(uint32_t* o, int n)
-{
-	uint32_t res = *o;
-	*o += n;
-	return res;
-}
-
-#define dtoh8o(a, o)	dtoh8a ((a) + _post_inc(&o, sizeof(uint8_t)))
-#define dtoh16o(a, o)	dtoh16a((a) + _post_inc(&o, sizeof(uint16_t)))
-#define dtoh32o(a, o)	dtoh32a((a) + _post_inc(&o, sizeof(uint32_t)))
-#define dtoh64o(a, o)	dtoh64a((a) + _post_inc(&o, sizeof(uint64_t)))
-
-
 /*
  * PTP strings ... if the size field is:
  * size 0  : "empty string" ... we interpret that as string with just \0 terminator, return 1
@@ -433,21 +329,6 @@ ptp_unpack_DI (PTPParams *params, const unsigned char* data, PTPDeviceInfo *di, 
 	return 1;
 }
 
-inline static void
-ptp_free_DI (PTPDeviceInfo *di) {
-	free (di->SerialNumber);
-	free (di->DeviceVersion);
-	free (di->Model);
-	free (di->Manufacturer);
-	free (di->ImageFormats);
-	free (di->CaptureFormats);
-	free (di->VendorExtensionDesc);
-	free (di->OperationsSupported);
-	free (di->EventsSupported);
-	free (di->DevicePropertiesSupported);
-	memset(di, 0, sizeof(*di));
-}
-
 /* EOS Device Info unpack */
 static inline int
 ptp_unpack_EOS_DI (PTPParams *params, const unsigned char* data, PTPCanonEOSDeviceInfo *di, unsigned int datalen)
@@ -461,14 +342,6 @@ ptp_unpack_EOS_DI (PTPParams *params, const unsigned char* data, PTPCanonEOSDevi
 	ptp_unpack_uint32_t_array(params, data, &offset, datalen, &di->unk, &di->unk_len);
 
 	return offset >= 16;
-}
-
-static inline void
-ptp_free_EOS_DI (PTPCanonEOSDeviceInfo *di)
-{
-	free (di->EventsSupported);
-	free (di->DevicePropertiesSupported);
-	free (di->unk);
 }
 
 /* ObjectHandles array pack/unpack */
