+++ camlibs/ptp2/ptpip.c
@@ -63,14 +63,10 @@
 
 #include "ptp.h"
 #include "ptp-private.h"
+#include "ptp-bugs.h"
 
 #define PTPIP_VERSION_MAJOR 0x0001
 #define PTPIP_VERSION_MINOR 0x0000
-
-#include "ptp.h"
-#include "ptp-bugs.h"
-
-#include "ptp-pack.c"
 
 #define ptpip_len		0
 #define ptpip_type		4
